Added a transaction cache which holds values that are only visable by the particular client.
Events are no longer received while I am in a transaction.

Modified isCached and the read cache calls so that they reflect the 
right behavior with the transaction cache and the cache while
transactions are in progress.

Modified the calls for adding to the cache to be aware of whether we
are in a transaction.