Merged with
    http://smi-protege.stanford.edu/repos/protege/protege-core/branches/2010-06-13-lock

Fixed a serious race condition in the transaction cache code.  Added
the ability to make transactions occur one at a time.  This is kind of
nice - it allows the robots to run without exceptions.  Rearranged the
order of the value and idle connection narrow frame store allowing
multi-threaded clients to go to the cache before possibly getting
blocked at the idle connection narrow frame store.   Allow multiple
threads in the same session to execute database code when the
session in question is not in a transaction.